pyproject.toml
[tool.poetry]
name = "ggst-api"
version = "0.0.6"
description = "Guilty Gear Strive API wrapper"
authors = ["UnknownMemory <jeremyschneider21@gmail.com>"]
packages = [
    {include = "GGST"}
]

[tool.poetry.dependencies]
python = "^3.9"
msgpack = "^1.0.5"
requests = "^2.28.2"
python-dotenv = "^0.20.0"
steam = "^1.4.4"
gevent = "21.8.0"
gevent-eventemitter = "2.1"
greenlet = "1.1.2"
protobuf = "3.20.3"
pycryptodomex = "3.16.0"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
black = "^22.12.0"
pylint = "^2.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
